Add entities
