+++ homeassistant/components/jvc_projector/sensor.py
@@ -61,6 +61,24 @@ SENSORS: tuple[JvcProjectorSensorDescription, ...] = (
         device_class=SensorDeviceClass.ENUM,
         entity_registry_enabled_default=False,
     ),
+    JvcProjectorSensorDescription(
+        key="hdr",
+        command=cmd.Hdr,
+        device_class=SensorDeviceClass.ENUM,
+        entity_registry_enabled_default=False,
+    ),
+    JvcProjectorSensorDescription(
+        key="hdr_processing",
+        command=cmd.HdrProcessing,
+        device_class=SensorDeviceClass.ENUM,
+        entity_registry_enabled_default=False,
+    ),
+    JvcProjectorSensorDescription(
+        key="picture_mode",
+        command=cmd.PictureMode,
+        device_class=SensorDeviceClass.ENUM,
+        entity_registry_enabled_default=False,
+    ),
 )
